[orchestration] Refactor into common parent package
