+++ src/main/scala/temple/builder/project/ProjectBuilder.scala
@@ -11,7 +11,7 @@ import temple.generate.FileSystem._
 import temple.generate.database.ast.Statement
 import temple.generate.database.{PostgresContext, PostgresGenerator}
 import temple.generate.docker.DockerfileGenerator
-import temple.generate.kube.KubernetesGenerator
+import temple.generate.orchestration.kube.KubernetesGenerator
 import temple.generate.metrics.grafana.ast.Datasource
 import temple.generate.metrics.grafana.{GrafanaDashboardConfigGenerator, GrafanaDashboardGenerator, GrafanaDatasourceConfigGenerator}
 import temple.generate.metrics.prometheus.PrometheusConfigGenerator
